+++ crates/slb-cli/tests/cli_tests.rs
@@ -1,6 +1,0 @@
-#[test]
-fn cli_tests() {
-    trycmd::TestCases::new()
-        // .case("README.md")
-        .case("tests/cmd/*.trycmd");
-}