+++ crates/slb/src/lib.rs
@@ -1,4 +1,0 @@
-mod letters_boxed;
-
-pub use letters_boxed::LettersBoxed;
-pub use letters_boxed::Shuffle;