+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - 🔧 chore(dependencies)-format dependencies and add metadata(pr [#31])
+- ♻️ refactor(main)-update default source directory path(pr [#32])
 
 ## [0.1.7] - 2025-03-31
 
@@ -114,6 +115,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#29]: https://github.com/jerus-org/slb/pull/29
 [#30]: https://github.com/jerus-org/slb/pull/30
 [#31]: https://github.com/jerus-org/slb/pull/31
+[#32]: https://github.com/jerus-org/slb/pull/32
 [Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.7...HEAD
 [0.1.7]: https://github.com/jerus-org/slb/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/jerus-org/slb/compare/v0.1.5...v0.1.6