+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - restructure project into a workspace(pr [#33])
+- Add-max-chain-length-option-for-solutions(pr [#37])
 
 ### Security
 
@@ -133,6 +134,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#33]: https://github.com/jerus-org/slb/pull/33
 [#35]: https://github.com/jerus-org/slb/pull/35
 [#36]: https://github.com/jerus-org/slb/pull/36
+[#37]: https://github.com/jerus-org/slb/pull/37
 [Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.8...HEAD
 [0.1.8]: https://github.com/jerus-org/slb/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/jerus-org/slb/compare/v0.1.6...v0.1.7