+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.9] - 2025-04-02
 
 ### Added
 
@@ -141,7 +141,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#38]: https://github.com/jerus-org/slb/pull/38
 [#39]: https://github.com/jerus-org/slb/pull/39
 [#40]: https://github.com/jerus-org/slb/pull/40
-[Unreleased]: https://github.com/jerus-org/slb/compare/v0.1.8...HEAD
+[0.1.9]: https://github.com/jerus-org/slb/compare/v0.1.8...v0.1.9
 [0.1.8]: https://github.com/jerus-org/slb/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/jerus-org/slb/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/jerus-org/slb/compare/v0.1.5...v0.1.6